The Kestrel-2's very, very, very first program!!  FIRST CODE!  YAY!  There is, however, a bit of a problem.  Although the code seems to run fine in that it flickers a handful of pixels on and off on the screen, it is destructive.  That is, the static image that is pre-loaded into video memory gets overwritten with either $0000 or $FFFF.  I do not yet have an explanation for this behavior.  Still, this is mad progress.
As a natural side-effect of using the Wishbone bus for the J1A CPU, I have discovered several changes to the memory fetch and store operations.  Fetches, it seems, can get along just fine with a single cycle -- that is, I no longer need to insert a NOP in front of @.  This is because the Wishbone interface runs at one half the CPU's clock (CPU is clocked at 25MHz, but runs at 12.5MIPS because of the bus wait-states).  However, I require _three_ machine cycles to complete a write.  The address and data buses aren't ready for the write until after the rising edge of sys_clk_i, which is exactly half a clock cycle after the CPU receives its store instruction.  When that happens, the RAMs see their write-enable asserted only during the wait-state, thus turning the write operation into a read operation during the second acknowledged cycle.  Whoops!  The fix, so far, is to change the implementation of ! to emit a write-no-op ($6020), which causes the write to happen with the address and data bus intact, followed by two DROP instructions to dispose of the values on the stack, per ANS Forth expectations.